+++ src/lib.rs
@@ -1,37 +1,38 @@
 //! # Overview
-//! Update informer for CLI applications written in Rust. It checks for a new version on **Crates.io**, **GitHub**, and **PyPI**.
+//! Update informer for CLI applications written in Rust. It checks for a new version on [`Crates.io`], [`GitHub`] and [`PyPI`].
 //!
-//! ### Benefits
-//! * Support of **Crates.io**, *GitHub**, and **PyPI**.
-//! * **Configurable frequency** of checks.
-//! * **Minimum dependencies** - only [ureq](https://github.com/algesten/ureq), [semver](https://github.com/dtolnay/semver) and [serde](https://github.com/serde-rs/serde).
+//! ## Benefits
+//! * Support of [`Crates.io`], [`GitHub`] and [`PyPI`].
+//! * Configurable check frequency and request timeout.
+//! * Minimum dependencies - only [`ureq`], [`semver`] and [`serde`].
 //!
 //! ## Usage
 //!
-//! To check for a new version on **Crates.io**, use the [UpdateInformer::check_version] function. This function takes the project name and current version as well as check interval:
+//! To check for a new version on Crates.io, use the [`UpdateInformer::check_version`] function. This function takes the project name and current version:
 //!
 //! ```rust
-//! use std::time::Duration;
 //! use update_informer::{registry::Crates, Check, UpdateInformer};
 //!
-//! let informer = UpdateInformer::new(Crates, "repo", "0.1.0", Duration::from_secs(60 * 60 * 24));
+//! let informer = UpdateInformer::new(Crates, "crate_name", "0.1.0");
 //! if let Ok(Some(version)) = informer.check_version() {
 //!     println!("New version is available: {}", version);
 //! }
 //! ```
 //!
-//! Also, you can take the name and version of the project from **Cargo** using environment variables:
+//! Also, you can take the name and version of the project from `Cargo` using environment variables:
 //!
 //! ```rust
-//! use std::time::Duration;
 //! use update_informer::{registry::Crates, Check, UpdateInformer};
 //!
 //! let name = env!("CARGO_PKG_NAME");
 //! let version = env!("CARGO_PKG_VERSION");
-//! UpdateInformer::new(Crates, name, version, Duration::from_secs(60 * 60 * 24)).check_version();
+//! UpdateInformer::new(Crates, name, version).check_version();
 //! ```
 //!
-//! Note that the first check will start only after the interval has expired:
+//! ## Interval
+//!
+//! Note that the first check will start only after the interval has expired.
+//! By default, the interval is 24 hours, but you can change it:
 //!
 //! ```rust
 //! use std::time::Duration;
@@ -39,56 +40,71 @@
 //!
 //! const EVERY_HOUR: Duration = Duration::from_secs(60 * 60);
 //!
-//! let informer = UpdateInformer::new(Crates, "repo", "0.1.0", EVERY_HOUR);
+//! let informer = UpdateInformer::new(Crates, "crate_name", "0.1.0").interval(EVERY_HOUR);
 //! informer.check_version(); // The check will start only after an hour
 //! ```
+//! ## Request timeout
 //!
-//! To check for a new version on **GitHub** (note that the project name must contain the owner):
+//! You can also change the request timeout. By default, it is 5 seconds:
 //!
 //! ```rust
 //! use std::time::Duration;
-//! use update_informer::{registry::GitHub, Check, UpdateInformer};
+//! use update_informer::{registry::Crates, Check, UpdateInformer};
 //!
-//! let informer = UpdateInformer::new(GitHub, "owner/repo", "0.1.0", Duration::from_secs(60 * 60 * 24));
+//! const THIRTY_SECONDS: Duration = Duration::from_secs(30);
+//!
+//! let informer = UpdateInformer::new(Crates, "crate_name", "0.1.0").timeout(THIRTY_SECONDS);
 //! informer.check_version();
 //! ```
 //!
-//! To check for a new version on **PyPI**:
+//! ## GitHub
+//!
+//! To check for a new version on GitHub (note that the project name must contain the owner):
 //!
 //! ```rust
-//! use std::time::Duration;
+//! use update_informer::{registry::GitHub, Check, UpdateInformer};
+//!
+//! let informer = UpdateInformer::new(GitHub, "owner/repo", "0.1.0");
+//! informer.check_version();
+//! ```
+//!
+//! ## PyPi
+//!
+//! To check for a new version on PyPI:
+//!
+//! ```rust
 //! use update_informer::{registry::PyPI, Check, UpdateInformer};
 //!
-//! let informer = UpdateInformer::new(PyPI, "yourpackage", "0.1.0", Duration::from_secs(60 * 60 * 24));
+//! let informer = UpdateInformer::new(PyPI, "package_name", "0.1.0");
 //! informer.check_version();
 //! ```
 //!
 //! ## Tests
 //!
-//! In order not to check for updates in tests, you can use the [FakeUpdateInformer::check_version] function, which returns the desired version.
+//! In order not to check for updates in tests, you can use the [`FakeUpdateInformer::check_version`] function, which returns the desired version.
 //!
 //! Example of usage in unit tests:
 //!
 //! ```rust
-//! use std::time::Duration;
 //! use update_informer::{registry::Crates, Check, FakeUpdateInformer, UpdateInformer};
 //!
-//! let name = "repo";
+//! let name = "crate_name";
 //! let version = "0.1.0";
-//! let interval = Duration::from_secs(60 * 60 * 24);
 //!
 //! #[cfg(not(test))]
-//! let informer = UpdateInformer::new(Crates, name, version, interval);
+//! let informer = UpdateInformer::new(Crates, name, version);
 //!
 //! #[cfg(test)]
-//! let informer = FakeUpdateInformer::new(Crates, name, version, interval, "1.0.0");
+//! let informer = FakeUpdateInformer::new(Crates, name, version, "1.0.0");
 //!
 //! if let Ok(Some(version)) = informer.check_version() {
 //!     println!("New version is available: {}", version);
 //! }
 //! ```
 //!
-//! To use the [FakeUpdateInformer::check_version] function in integration tests, you must first add the feature flag to `Cargo.toml`:
+//! ## Integration tests
+//!
+//! To use the [`FakeUpdateInformer::check_version`] function in integration tests, you must first add the feature flag to `Cargo.toml`:
 //!
 //! ```toml
 //! [features]
@@ -98,21 +114,26 @@
 //! Then use this feature flag in your code and integration tests:
 //!
 //! ```rust
-//! use std::time::Duration;
 //! use update_informer::{registry::Crates, Check, FakeUpdateInformer, UpdateInformer};
 //!
-//! let name = "repo";
+//! let name = "crate_name";
 //! let version = "0.1.0";
-//! let interval = Duration::from_secs(60 * 60 * 24);
 //!
 //! #[cfg(not(feature = "stub_check_version"))]
-//! let informer = UpdateInformer::new(Crates, name, version, interval);
+//! let informer = UpdateInformer::new(Crates, name, version);
 //!
 //! #[cfg(feature = "stub_check_version")]
-//! let informer = FakeUpdateInformer::new(Crates, name, version, interval, "1.0.0");
+//! let informer = FakeUpdateInformer::new(Crates, name, version, "1.0.0");
 //!
 //! informer.check_version();
 //! ```
+//!
+//! [`Crates.io`]: https://crates.io
+//! [`GitHub`]: https://github.com
+//! [`PyPI`]: https://pypi.org
+//! [`ureq`]: https://github.com/algesten/ureq
+//! [`semver`]: https://github.com/dtolnay/semver
+//! [`serde`]: https://github.com/serde-rs/serde
 
 use crate::package::Package;
 use crate::registry::Registry;
@@ -120,6 +141,7 @@ use crate::version::Version;
 use crate::version_file::VersionFile;
 use std::time::Duration;
 
+mod http;
 mod package;
 mod version;
 mod version_file;
@@ -139,12 +161,13 @@ pub trait Check {
     }
 }
 
-/// Checks for a new version on Crates.io and GitHub.
+/// Checks for a new version on Crates.io, GitHub and PyPi
 pub struct UpdateInformer<R: Registry, N: AsRef<str>, V: AsRef<str>> {
     _registry: R,
     name: N,
     version: V,
     interval: Duration,
+    timeout: Duration,
 }
 
 impl<R: Registry, N: AsRef<str>, V: AsRef<str>> UpdateInformer<R, N, V> {
@@ -155,7 +178,31 @@ impl<R: Registry, N: AsRef<str>, V: AsRef<str>> UpdateInformer<R, N, V> {
     /// * `registry` - A registry service such as Crates.io or GitHub.
     /// * `name` - A project name.
     /// * `version` - Current version of the project.
-    /// * `interval` - An interval how often to check for a new version.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use update_informer::{registry::Crates, Check, UpdateInformer};
+    ///
+    /// let name = env!("CARGO_PKG_NAME");
+    /// let version = env!("CARGO_PKG_VERSION");
+    /// let informer = UpdateInformer::new(Crates, name, version);
+    /// ```
+    pub fn new(registry: R, name: N, version: V) -> Self {
+        Self {
+            _registry: registry,
+            name,
+            version,
+            interval: Duration::from_secs(60 * 60 * 24), // Once a day
+            timeout: Duration::from_secs(5),
+        }
+    }
+
+    /// Sets an interval how often to check for a new version.
+    ///
+    /// # Arguments
+    ///
+    /// * `interval` - An interval in seconds. By default, it is 24 hours.
     ///
     /// # Examples
     ///
@@ -163,17 +210,34 @@ impl<R: Registry, N: AsRef<str>, V: AsRef<str>> UpdateInformer<R, N, V> {
     /// use std::time::Duration;
     /// use update_informer::{registry::Crates, Check, UpdateInformer};
     ///
-    /// let name = env!("CARGO_PKG_NAME");
-    /// let version = env!("CARGO_PKG_VERSION");
-    /// let informer = UpdateInformer::new(Crates, name, version, Duration::from_secs(60 * 60 * 24));
+    /// const EVERY_HOUR: Duration = Duration::from_secs(60 * 60);
+    ///
+    /// let informer = UpdateInformer::new(Crates, "crate_name", "0.1.0").interval(EVERY_HOUR);
+    /// informer.check_version(); // The check will start only after an hour
     /// ```
-    pub fn new(registry: R, name: N, version: V, interval: Duration) -> Self {
-        Self {
-            _registry: registry,
-            name,
-            version,
-            interval,
-        }
+    pub fn interval(self, interval: Duration) -> Self {
+        Self { interval, ..self }
+    }
+
+    /// Sets a request timeout.
+    ///
+    /// # Arguments
+    ///
+    /// * `timeout` - A request timeout. By default, it is 5 seconds.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use std::time::Duration;
+    /// use update_informer::{registry::Crates, Check, UpdateInformer};
+    ///
+    /// const THIRTY_SECONDS: Duration = Duration::from_secs(30);
+    ///
+    /// let informer = UpdateInformer::new(Crates, "crate_name", "0.1.0").timeout(THIRTY_SECONDS);
+    /// informer.check_version();
+    /// ```
+    pub fn timeout(self, timeout: Duration) -> Self {
+        Self { timeout, ..self }
     }
 }
 
@@ -182,35 +246,13 @@ impl<R: Registry, N: AsRef<str>, V: AsRef<str>> Check for UpdateInformer<R, N, V
     ///
     /// # Examples
     ///
-    /// To check for a new version on **Crates.io**:
+    /// To check for a new version on Crates.io:
     ///
     /// ```rust
-    /// use std::time::Duration;
     /// use update_informer::{registry::Crates, Check, UpdateInformer};
     ///
-    /// let informer = UpdateInformer::new(Crates, "repo", "0.1.0", Duration::from_secs(60 * 60 * 24));
+    /// let informer = UpdateInformer::new(Crates, "crate_name", "0.1.0");
     /// informer.check_version();
-    /// ```
-    ///
-    /// To check for a new version on **GitHub**:
-    ///
-    /// ```rust
-    /// use std::time::Duration;
-    /// use update_informer::{registry::GitHub, Check, UpdateInformer};
-    ///
-    /// let informer = UpdateInformer::new(GitHub, "owner/repo", "0.1.0", Duration::from_secs(60 * 60 * 24));
-    /// informer.check_version();
-    /// ```
-    /// Note that the first check will start only after the interval has expired:
-    ///
-    /// ```rust
-    /// use std::time::Duration;
-    /// use update_informer::{registry::Crates, Check, UpdateInformer};
-    ///
-    /// const EVERY_HOUR: Duration = Duration::from_secs(60 * 60);
-    ///
-    /// let informer = UpdateInformer::new(Crates, "repo", "0.1.0", EVERY_HOUR);
-    /// informer.check_version(); // The check will start only after an hour
     /// ```
     fn check_version(&self) -> Result<Option<Version>, Error> {
         let pkg = Package::new(self.name.as_ref());
@@ -221,7 +263,7 @@ impl<R: Registry, N: AsRef<str>, V: AsRef<str>> Check for UpdateInformer<R, N, V
             // This is needed to update mtime of the file
             latest_version_file.recreate_file()?;
 
-            match R::get_latest_version(&pkg)? {
+            match R::get_latest_version(&pkg, self.timeout)? {
                 Some(v) => {
                     latest_version_file.write_version(&v)?;
                     v
@@ -261,18 +303,11 @@ impl<V: AsRef<str>> FakeUpdateInformer<V> {
     /// # Examples
     ///
     /// ```rust
-    /// use std::time::Duration;
     /// use update_informer::{registry::Crates, Check, FakeUpdateInformer};
     ///
-    /// let informer = FakeUpdateInformer::new(Crates, "repo", "0.1.0", Duration::from_secs(60 * 60 * 24), "1.0.0");
+    /// let informer = FakeUpdateInformer::new(Crates, "repo", "0.1.0", "1.0.0");
     /// ```
-    pub fn new<R, N>(
-        _registry: R,
-        _name: N,
-        _version: V,
-        _interval: Duration,
-        new_version: V,
-    ) -> Self
+    pub fn new<R, N>(_registry: R, _name: N, _version: V, new_version: V) -> Self
     where
         R: Registry,
         N: AsRef<str>,
@@ -282,6 +317,16 @@ impl<V: AsRef<str>> FakeUpdateInformer<V> {
             version: new_version,
         }
     }
+
+    /// Returns `FakeUpdateInformer`.
+    pub fn interval(self, _interval: Duration) -> Self {
+        self
+    }
+
+    /// Returns `FakeUpdateInformer`.
+    pub fn timeout(self, _timeout: Duration) -> Self {
+        self
+    }
 }
 
 impl<V: AsRef<str>> Check for FakeUpdateInformer<V> {
@@ -290,10 +335,9 @@ impl<V: AsRef<str>> Check for FakeUpdateInformer<V> {
     /// # Examples
     ///
     /// ```rust
-    /// use std::time::Duration;
     /// use update_informer::{registry::Crates, Check, FakeUpdateInformer};
     ///
-    /// let informer = FakeUpdateInformer::new(Crates, "repo", "0.1.0", Duration::from_secs(60 * 60 * 24), "1.0.0");
+    /// let informer = FakeUpdateInformer::new(Crates, "crate_name", "0.1.0", "1.0.0");
     /// let result = informer.check_version();
     /// assert!(result.is_ok());
     ///
@@ -319,7 +363,6 @@ mod tests {
     const PKG_NAME: &str = "repo";
     const CURRENT_VERSION: &str = "3.1.0";
     const LATEST_VERSION: &str = "3.1.1";
-    const ONE_DAY: Duration = Duration::from_secs(60 * 60 * 24);
 
     fn mock_github(pkg: &str) -> Mock {
         let pkg = Package::new(pkg);
@@ -346,7 +389,7 @@ mod tests {
     #[test]
     fn no_new_version_with_interval_test() {
         within_test_dir(|_| {
-            let informer = UpdateInformer::new(Crates, PKG_NAME, CURRENT_VERSION, ONE_DAY);
+            let informer = UpdateInformer::new(Crates, PKG_NAME, CURRENT_VERSION);
             let result = informer.check_version();
 
             assert!(result.is_ok());
@@ -359,7 +402,7 @@ mod tests {
         within_test_dir(|_| {
             let _mock = mock_crates(PKG_NAME);
             let informer =
-                UpdateInformer::new(Crates, PKG_NAME, LATEST_VERSION, Duration::default());
+                UpdateInformer::new(Crates, PKG_NAME, LATEST_VERSION).interval(Duration::default());
             let result = informer.check_version();
 
             assert!(result.is_ok());
@@ -371,8 +414,8 @@ mod tests {
     fn check_version_on_github_test() {
         within_test_dir(|_| {
             let _mock = mock_github("owner/repo");
-            let informer =
-                UpdateInformer::new(GitHub, "owner/repo", CURRENT_VERSION, Duration::default());
+            let informer = UpdateInformer::new(GitHub, "owner/repo", CURRENT_VERSION)
+                .interval(Duration::default());
             let result = informer.check_version();
             let version = Version::parse(LATEST_VERSION).expect("parse version");
 
@@ -385,8 +428,8 @@ mod tests {
     fn check_version_on_crates_test() {
         within_test_dir(|_| {
             let _mock = mock_crates(PKG_NAME);
-            let informer =
-                UpdateInformer::new(Crates, PKG_NAME, CURRENT_VERSION, Duration::default());
+            let informer = UpdateInformer::new(Crates, PKG_NAME, CURRENT_VERSION)
+                .interval(Duration::default());
             let result = informer.check_version();
             let version = Version::parse(LATEST_VERSION).expect("parse version");
 
@@ -400,7 +443,7 @@ mod tests {
         within_test_dir(|version_file| {
             fs::write(version_file, "4.0.0").expect("create file");
 
-            let informer = UpdateInformer::new(Crates, PKG_NAME, CURRENT_VERSION, ONE_DAY);
+            let informer = UpdateInformer::new(Crates, PKG_NAME, CURRENT_VERSION);
             let result = informer.check_version();
             let version = Version::parse("4.0.0").expect("parse version");
 
@@ -414,7 +457,7 @@ mod tests {
         within_test_dir(|version_file| {
             assert!(!version_file.exists());
 
-            let informer = UpdateInformer::new(Crates, PKG_NAME, CURRENT_VERSION, ONE_DAY);
+            let informer = UpdateInformer::new(Crates, PKG_NAME, CURRENT_VERSION);
             let result = informer.check_version();
             assert!(result.is_ok());
             assert!(version_file.exists());
@@ -428,7 +471,7 @@ mod tests {
     fn check_version_with_string_name_test() {
         within_test_dir(|_| {
             let pkg_name = format!("{}/{}", "owner", PKG_NAME);
-            let informer = UpdateInformer::new(Crates, pkg_name, CURRENT_VERSION, ONE_DAY);
+            let informer = UpdateInformer::new(Crates, pkg_name, CURRENT_VERSION);
             let result = informer.check_version();
 
             assert!(result.is_ok());
@@ -439,7 +482,7 @@ mod tests {
     fn check_version_with_string_version_test() {
         within_test_dir(|_| {
             let version = String::from(CURRENT_VERSION);
-            let informer = UpdateInformer::new(Crates, PKG_NAME, version, ONE_DAY);
+            let informer = UpdateInformer::new(Crates, PKG_NAME, version);
             let result = informer.check_version();
 
             assert!(result.is_ok());
@@ -451,7 +494,7 @@ mod tests {
         within_test_dir(|_| {
             let pkg_name = format!("{}/{}", "owner", PKG_NAME);
             let version = String::from(CURRENT_VERSION);
-            let informer = UpdateInformer::new(Crates, &pkg_name, &version, ONE_DAY);
+            let informer = UpdateInformer::new(Crates, &pkg_name, &version);
             let result = informer.check_version();
 
             assert!(result.is_ok());
@@ -461,7 +504,7 @@ mod tests {
     #[test]
     fn fake_check_version_test() {
         let version = "1.0.0";
-        let informer = FakeUpdateInformer::new(Crates, PKG_NAME, CURRENT_VERSION, ONE_DAY, version);
+        let informer = FakeUpdateInformer::new(Crates, PKG_NAME, CURRENT_VERSION, version);
         let result = informer.check_version();
         let version = Version::parse(version).expect("parse version");