+++ src/registry/pypi.rs
@@ -1,6 +1,7 @@
 use crate::registry::Registry;
-use crate::{Error, Package};
+use crate::{http, Error, Package};
 use serde::Deserialize;
+use std::time::Duration;
 
 #[cfg(test)]
 use mockito;
@@ -32,10 +33,10 @@ fn get_base_url() -> String {
 impl Registry for PyPI {
     const NAME: &'static str = "pypi";
 
-    fn get_latest_version(pkg: &Package) -> Result<Option<String>, Error> {
+    fn get_latest_version(pkg: &Package, timeout: Duration) -> Result<Option<String>, Error> {
         let url = format!("{}/{}/json", get_base_url(), pkg);
 
-        let resp: Response = ureq::get(&url).call()?.into_json()?;
+        let resp: Response = http::get(&url, timeout).call()?;
 
         if !resp.info.yanked {
             return Ok(Some(resp.info.version));
@@ -52,6 +53,7 @@ mod tests {
 
     const PKG_NAME: &str = "filprofiler";
     const FIXTURES_PATH: &str = "tests/fixtures/registry/pypi";
+    const TIMEOUT: Duration = Duration::from_secs(5);
 
     #[test]
     fn failure_test() {
@@ -59,7 +61,7 @@ mod tests {
         let data_path = format!("{}/not_found.json", FIXTURES_PATH);
         let _mock = mock_pypi(&pkg, 404, &data_path);
 
-        let result = PyPI::get_latest_version(&pkg);
+        let result = PyPI::get_latest_version(&pkg, TIMEOUT);
         assert!(result.is_err());
     }
 
@@ -70,7 +72,7 @@ mod tests {
         let (_mock, _data) = mock_pypi(&pkg, 200, &data_path);
 
         let latest_version = "2022.1.1".to_string();
-        let result = PyPI::get_latest_version(&pkg);
+        let result = PyPI::get_latest_version(&pkg, TIMEOUT);
 
         assert!(result.is_ok());
         assert_eq!(result.expect("get result"), Some(latest_version));